add process service test
